pyproject.toml
[project]
name = "winvite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.12.0",
    "psycopg[binary,pool]>=3.1",
    "polars>=1.0",
    "python-dotenv>=1.0",
]
